Revamp swim AI and class hierarchy of fish
