+++ src/main/java/com/sirjain/entities/entity/MaxillaMortisEntity.java
@@ -38,8 +38,8 @@ public class MaxillaMortisEntity extends NoBucketFishEntity {
 	public MaxillaMortisEntity(EntityType<? extends FishEntity> entityType, World world) {
 		super(entityType, world);
 		this.setTargetState(false);
-		this.moveControl = new AquaticMoveControl(this, 20, 10, 1, 0.04F, true);
-		this.lookControl = new YawAdjustingLookControl(this, 10);
+//		this.moveControl = new AquaticMoveControl(this, 20, 10, 1, 0.04F, true);
+//		this.lookControl = new YawAdjustingLookControl(this, 10);
 	}
 
 	@Override
@@ -67,10 +67,10 @@ public class MaxillaMortisEntity extends NoBucketFishEntity {
 		this.targetSelector.add(1, new RevengeGoal(this));
 	}
 
-	@Override
-	protected EntityNavigation createNavigation(World world) {
-		return new SwimNavigation(this, world);
-	}
+//	@Override
+//	protected EntityNavigation createNavigation(World world) {
+//		return new SwimNavigation(this, world);
+//	}
 
 	@Override
 	protected void initDataTracker() {
@@ -110,12 +110,6 @@ public class MaxillaMortisEntity extends NoBucketFishEntity {
 
 		if (name != null && name.equals(Text.literal("Fibula")))
 			this.setCanBeRecolored(true);
-	}
-
-	// TODO: Change to generic fish flop sound
-	@Override
-	protected SoundEvent getFlopSound() {
-		return SoundEvents.ENTITY_COD_FLOP;
 	}
 
 	@Override
