+++ src/main/java/com/sirjain/entities/entity/template/NoBucketFishEntity.java
@@ -10,21 +10,12 @@ import net.minecraft.util.ActionResult;
 import net.minecraft.util.Hand;
 import net.minecraft.world.World;
 
-public abstract class NoBucketFishEntity extends FishEntity {
+public abstract class NoBucketFishEntity extends APFishEntity {
 	public NoBucketFishEntity(EntityType<? extends FishEntity> entityType, World world) {
 		super(entityType, world);
 	}
 
-	@Override
-	protected SoundEvent getFlopSound() {
-		return SoundEvents.ENTITY_COD_FLOP;
-	}
-
-	@Override
-	public ItemStack getBucketItem() {
-		return ItemStack.EMPTY;
-	}
-
+	// Returns a pass so that nothing happens when right clicking fish
 	@Override
 	protected ActionResult interactMob(PlayerEntity player, Hand hand) {
 		return ActionResult.PASS;
